Refactor: 서비스 로직 정리
