+++ src/service/video.ts
@@ -1,0 +1,7 @@
+export const extractYouTubeVideoId = (url: string) => {
+  const regex =
+    /(?:https?:\/\/)?(?:www\.)?(?:youtube\.com\/(?:[^\/\n\s]+\/\S+\/|(?:v|e(?:mbed)?)\/|\S*?[?&]v=)|youtu\.be\/)([a-zA-Z0-9_-]{11})/;
+  const matches = url.match(regex);
+
+  return matches ? matches[1] : null;
+};
